# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer Service ",
        "version": "0.1"
    },
    "paths": {
        "/api/ddq-groups": {
            "get": {
                "tags": [
                    "Customer DDQ Groups"
                ],
                "summary": "Get a list of DDQ Groups",
                "description": "Fetches a list of all available DDQ groups.",
                "operationId": "b6a08250aa114014b69aeb64586238c6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CustomerDdqGroup"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Customer DDQ Groups"
                ],
                "summary": "Create a new DDQ Group",
                "description": "Creates a new DDQ group with the provided title and status.",
                "operationId": "fa87ce8d08a7a63bff96d03ba1b78fdf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Group"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "DDQ Group created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "": {
                                            "$ref": "#/components/schemas/CustomerDdqGroup"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/ddq-groups/{id}": {
            "get": {
                "tags": [
                    "Customer DDQ Groups"
                ],
                "summary": "Get a specific DDQ Group by ID",
                "description": "Fetches a DDQ group by its unique ID.",
                "operationId": "8c1679f6445a954122fd0a41245c217b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the DDQ Group",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "": {
                                            "$ref": "#/components/schemas/CustomerDdqGroup"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customer DDQ Groups"
                ],
                "summary": "Update a specific DDQ Group",
                "description": "Updates a DDQ group with the specified ID using the provided data.",
                "operationId": "5db1f4dc8d9f1598521a38629bbe1ece",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the DDQ Group",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Group"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "DDQ Group updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "": {
                                            "$ref": "#/components/schemas/CustomerDdqGroup"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customer DDQ Groups"
                ],
                "summary": "Delete a specific DDQ Group",
                "description": "Deletes the DDQ group with the specified ID.",
                "operationId": "ad8dee8621c4cb18939db299d6150b2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the DDQ Group",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "DDQ Group deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerDdqGroup": {
                "title": "Customer DDQ Group",
                "properties": {
                    "id": {
                        "description": "ID of the DDQ Group",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title of the DDQ Group",
                        "type": "string",
                        "example": "Group Title"
                    },
                    "status": {
                        "description": "Status of the DDQ Group",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Creation date of the DDQ Group",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update date of the DDQ Group",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}